[windows] Fix std::stringstream error after upgrading libc++
Upgrading libc++ adds a deleted overload for operator<< when writing a
wchar_t* to a std::stringstream, because it surprisingly prints an
address rather than the string content. See the proposal in
wg21.link/p1423r3. Converting to a UTF-8 string is tedious, and I don't
know how to test this code, so make the surprising behavior explicit to
fix the compiler error.

Bug: 175635923
Test: m adb adbd
Change-Id: I50ee7087d2e8fcd6a7dae9c8ffb99563be34051b