Hide the impl detail of command verification tests
Hide the impl detail of how to get the atest internal command, i.e. the call of 'atest-dev -g'. The current impl of the verify_command function uses the runner_commands.json to get the atest internal command, but this logic is spread both inside and outside of the function (the reading of the runner_command.json is inside the function, the use of -g and reading json dictionary is outside of the function). This CL consolidate all the impl detail of getting the atest internal command into the verify_command function itself so that the test cases can focus on checking the actual command components.

Test: go/run-atest-integration-tests
Change-Id: I1378db0a8bb2469d3e13b9128fdcfc47eb32457a