tp: implement dominator tree intrinsic and corresponding macro
This CL adds the implementation of a "dominator tree" macro to the
standard library, backed by a new table function intrinsic in C++
implementing the Lengauer-Tarjan algorithm to compute the dominator tree
from a directed flow-graph.

This CL also brings a new approach to how we can make use of table
function infrastructure to operate on full tables by making use of the
`RepeatedField` function. This can potentially be even further improved
in the future but for now, this makes the user-facing API very nice.
This approach should be used on upcoming work like interval_intersect/
interval_set_difference etc.

Change-Id: Iecdc6a4b7eccc76ff46e0681bda9d16b745d2614