avbtool: open hash image read-only when possible
The `add_hash_footer` command normally adds the hash descriptor into the
image directly, except when the `--do_not_append_vbmeta_image` flag is
given. In this case we don't touch the input image so can open it
read-only.

An example use case is:
  1. Android source tree is checked out read-only (e.g. infra bots)
  2. Some test image is checked into the tree as a binary
  3. We want to call `add_hash_footer` on test image as a build rule
In this case, `avbtool` will fail during build if it attempts to open
the read-only test image as writable, so must open it read-only.

Bug: b/324451620
Test: atest libavb_host_unittest
Change-Id: I438030fed51244769100f305fcec7116fc679f4f