Autoupdate of 3 submodules
*external/icu updated to 603c71a7dc1e250db75b026ff28d0b7042f2693a
Remove unnecessary final modifier from enum methods

Metalava removes `final` modifier from methods in `final` classes to
normalize the model, and ensure consistency as historically the
underlying Java/Kotlin parsing code was inconsistent in how it
handled implicit or inherited modifiers. Unfortunately, due to a bug
in Metalava that normalization was not applied to methods in an `enum`
class (which is itself implicitly `final`). That bug has been fixed
and this is updating the few places where that affected these signature
files.

Bug: 326310840
Test: m checkapi
Change-Id: Ie1b6486b0416f0f2fcb1430ec11da916c671bf56

*libcore updated to 7060a59577e986c21b6c8353174f232c3bdfdbba
Merge "Remove unnecessary final modifier from enum methods" into main
*tools/metalava updated to 622cd506326c7fd96d44eb3c9189008dc8e8eb97
Clear final on enum methods in a final class

Previously, it would clear the `final` modifier on methods that are
not "in an enum" and not "an enum synthetic method" but it should have
cleared it on methods that are not "in an enum and are synthetic". As
the `isEnumSyntheticMethod()` already checks to make sure it is in an
enum class that can be simplified to just `!isEnumSyntheticMethod()`.
This change corrects that and updates the affected tests.

This will have a minor impact on API signature files downstream.

Bug: 326310840
Test: m checkapi
Change-Id: Ie00e092562f2d9960f579e17300a955fb9a5d857