Autoupdate of 6 submodules
*external/crosvm updated to 6cfac38770632902737cbd74e839fece6a246b08
Enable GPU feature

crosvm in virt APEX now supports GPU feature. This is to support VMs
with display.

Bug: 325929966
Test: m
Change-Id: I307d5eca3856ad60b722f1020f9f7a1ae6bd386c

*external/libdrm updated to 2d533abeaa34dbdc21b031ce39279d324b220017
Enable GPU in crosvm

When the GPU feature is turned on in crosvm, these modules are added as
dependencies. Since crosvm is included in the virt APEX, add the APEX to
the apex_available properties of the modules to make them available in
the APEX.

Bug: 325929966
Test: m
Change-Id: I680d5231895d045f94ad4d6ef559cafa08e4340e

*external/libyuv updated to 9a34347b9f5d1991bd0e19b56b88901eb4b0f858
Merge "Enable GPU in crosvm" into main
*external/lz4 updated to 793703b901c9cf989a723d567a740ab3322142e1
Enable GPU in crosvm

When the GPU feature is turned on in crosvm, these modules are added as
dependencies. Since crosvm is included in the virt APEX, add the APEX to
the apex_available properties of the modules to make them available in
the APEX.

Bug: 325929966
Test: m
Change-Id: If1435570fd5697af3b9989f0477014452456acbc

*hardware/google/gfxstream updated to ac97b7ae377528c099d3a470c147cc76fb380e21
Merge "Enable GPU in crosvm" into main
*packages/modules/Virtualization updated to 2e98792017b2732fef611e13b1a13d9293da7a4b
Merge "Enable GPU in crosvm" into main